Small fix regarding weird state with blobs
--HG--
extra : convert_revision : svn%3Ae08d0fb5-4147-0410-a205-bba44f1f51a3/trunk%40296

+++ bookworm/library/models.py
@@ -555,11 +555,15 @@ class ImageFile(BookwormFile):
 
     def delete(self):
         b = self._blob()
-        b.delete()
+        if b:
+            b.delete()
         super(ImageFile, self).save()
 
     def _blob(self):
         '''Gets the blob related to this image'''
+        b = self._blob_class().objects.filter(image=self)
+        if len(b) == 0:
+            return None
         return self._blob_class().objects.filter(image=self)[0]        
 
     def _blob_class(self):
